refactor(search): split search source helpers
Separate bbox slicing, vertical text reconstruction, and form or annotation source matching from the main search orchestration. This keeps each search module focused enough to review independently while preserving the existing match cap and duplicate suppression behavior.
